An FP-Tree structure that performs frequency analysis was created using the FP-Growth algorithm with grocery receipt data. Frequent item clusters are determined according to the minimum support value and displayed in a tree structure.
This code identifies frequent itemsets with the FP-Growth algorithm using data from a grocery receipt. How it works:

Data Loading: Imports voucher data from the CSV file and organizes the first 20 rows as a transaction set.

TreeNode Class: Creates FP-Tree nodes. Each node contains the item name, counter, and connections to other nodes.

FP-Tree Generation (fpTree): Frequent itemsets are determined based on the minimum support value and the header table is updated with FP-Tree.

Tree Update: Transactions are added or their counters are incremented with UpdateAgac; With UpdateTable, nodes of the same items are chained.

Result Display: The user creates the FP-Tree by entering the minimum support value and can see the tree structure on the screen. It can examine the relevant node by searching for a specific set of objects.

This code summarizes the data and organizes frequent items and links in a tree structure.